+++ CHANGELOG.md
@@ -5,6 +5,7 @@ PDF417 Changelog
 -----------
 
 * Added validation of options in renderers.
+* Fixed a bug in ImageRenderer where padding was not in bgColor, but white.
 
 0.1.0 (2014-12-24)
 ------------------